+++ linuxdoom-1.10/r_draw.c
@@ -27,6 +27,7 @@
 static const char
 rcsid[] = "$Id: r_draw.c,v 1.4 1997/02/03 16:47:55 b1 Exp $";
 
+#include <stdint.h>
 
 #include "doomdef.h"
 
@@ -461,7 +462,9 @@ void R_InitTranslationTables (void)
     int		i;
 	
     translationtables = Z_Malloc (256*3+255, PU_STATIC, 0);
-    translationtables = (byte *)(( (int)translationtables + 255 )& ~255);
+    // See R_InitColormaps in r_data.c: must align via a pointer-sized
+    // integer, not int, or this truncates the pointer on a 64-bit host.
+    translationtables = (byte *)(( (uintptr_t)translationtables + 255 )& ~(uintptr_t)255);
     
     // translate just the 16 green colors
     for (i=0 ; i<256 ; i++)
